[GraphQL/Package Resolver] E2E tests for type resolution limits
Add tests that the limits implemented in the package resolver help to
protect the GraphQL service against malicious type queries.

Note that tests were not added for the `max_type_nodes` and
`max_type_argument_width` limits because the RPC limits are the same
as the chain limits in these cases, so it's not possible to generate a
type that could hit these limits in RPC, because they are already
stopped by the chain's protocol config.  These limits have been
separately tested in the package resolver crate.

Test Plan:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- types.move
```